Breno branch

+++ demo/src/main/java/com/rtcc/demo/controller/CourseController.java
@@ -2,8 +2,10 @@ package com.rtcc.demo.controller;
 
 import com.rtcc.demo.DTOs.CourseRequestDTO;
 import com.rtcc.demo.DTOs.CourseResponseDTO;
+import com.rtcc.demo.exception.EntityDeletionException;
 import com.rtcc.demo.model.Course;
 import com.rtcc.demo.repository.CourseRepository;
+import com.rtcc.demo.repository.TccRepository;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -24,6 +26,9 @@ public class CourseController {
 
     @Autowired
     private CourseRepository courseRepository;
+
+    @Autowired
+    private TccRepository tccRepository;
 
 
     @Operation(summary = "Save a new course",
@@ -91,8 +96,14 @@ public class CourseController {
         if (!courseRepository.existsById(id)) {
             return ResponseEntity.notFound().build();
         }
+
+        boolean isLinkedToTcc = tccRepository.existsByCourseId(id);
+        if (isLinkedToTcc) {
+            throw new EntityDeletionException("Curso", "Não pode deletar um curso vinculado a um ou vários TCC");
+        }
+
         courseRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
